+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 ### v0.1.11
-* Fix Run.start_time and Run.end_time testing [ ]
+* Fix Run.start_time and Run.end_time testing [x]
 * Update README [ ]
 * Add subpackage docstrings [ ]
 * TBD [ ]